// sanity/types.ts
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export type BERITA_LIST_QUERYResult = Array<{
  _id: string;
  judul: string | null;
  slug: Slug | null;
  publishedAt: string | null;
}>;